Fix IP string Concatenation  & commented out code
The _ip that was originating from the legacy code had the string Concatenation done wrong. (how the f this worked...) . Moreover the functions that made the connection to the instrument were left commented out possibly from debuging.

Note: Noticed some issues in the dataloging form , no data was showing in the chart but we saw it in the csv